net, admitter: Move binding validation combinations to its own file
Binding tests have been relocated inside the `network/admitter` package.

Additional validations are to be added in follow-up changes from the
`vmi-create-admitter`.
