+++ pkg/network/admitter/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "admit.go",
+        "binding.go",
         "slirp.go",
         "validator.go",
     ],
@@ -22,6 +23,7 @@ go_test(
     srcs = [
         "admit_suite_test.go",
         "admit_test.go",
+        "binding_test.go",
         "slirp_test.go",
     ],
     deps = [
