+++ pkg/network/admitter/admit.go
@@ -59,31 +59,6 @@ func validateInterfaceStateValue(field *k8sfield.Path, spec *v1.VirtualMachineIn
 	return causes
 }
 
-func validateInterfaceBinding(field *k8sfield.Path, spec *v1.VirtualMachineInstanceSpec) []metav1.StatusCause {
-	var causes []metav1.StatusCause
-	for idx, iface := range spec.Domain.Devices.Interfaces {
-		if iface.Binding != nil {
-			if hasInterfaceBindingMethod(iface) {
-				causes = append(causes, metav1.StatusCause{
-					Type:    metav1.CauseTypeFieldValueInvalid,
-					Message: fmt.Sprintf("logical %s interface cannot have both binding plugin and interface binding method", iface.Name),
-					Field:   field.Child("domain", "devices", "interfaces").Index(idx).Child("binding").String(),
-				})
-			}
-		}
-	}
-	return causes
-}
-
-func hasInterfaceBindingMethod(iface v1.Interface) bool {
-	return iface.InterfaceBindingMethod.Bridge != nil ||
-		iface.InterfaceBindingMethod.Slirp != nil ||
-		iface.InterfaceBindingMethod.Masquerade != nil ||
-		iface.InterfaceBindingMethod.SRIOV != nil ||
-		iface.InterfaceBindingMethod.Macvtap != nil ||
-		iface.InterfaceBindingMethod.Passt != nil
-}
-
 func validateSinglePodNetwork(field *k8sfield.Path, spec *v1.VirtualMachineInstanceSpec) []metav1.StatusCause {
 	if countPodNetworks(spec.Networks) > 1 {
 		return []metav1.StatusCause{{
