+++ pkg/network/admitter/admit_test.go
@@ -94,42 +94,6 @@ var _ = Describe("Validating VMI network spec", func() {
 			}))
 	})
 
-	It("network interface has both binding plugin and interface binding method", func() {
-		vm := api.NewMinimalVMI("testvm")
-		vm.Spec.Domain.Devices.Interfaces = []v1.Interface{{
-			Name:                   "foo",
-			InterfaceBindingMethod: v1.InterfaceBindingMethod{Bridge: &v1.InterfaceBridge{}},
-			Binding:                &v1.PluginBinding{Name: "boo"},
-		}}
-		validator := admitter.NewValidator(k8sfield.NewPath("fake"), &vm.Spec, stubSlirpClusterConfigChecker{})
-		Expect(validator.Validate()).To(
-			ConsistOf(metav1.StatusCause{
-				Type:    "FieldValueInvalid",
-				Message: "logical foo interface cannot have both binding plugin and interface binding method",
-				Field:   "fake.domain.devices.interfaces[0].binding",
-			}))
-	})
-
-	It("network interface has only plugin binding", func() {
-		vm := api.NewMinimalVMI("testvm")
-		vm.Spec.Domain.Devices.Interfaces = []v1.Interface{{
-			Name:    "foo",
-			Binding: &v1.PluginBinding{Name: "boo"},
-		}}
-		validator := admitter.NewValidator(k8sfield.NewPath("fake"), &vm.Spec, stubSlirpClusterConfigChecker{})
-		Expect(validator.Validate()).To(BeEmpty())
-	})
-
-	It("network interface has only binding method", func() {
-		vm := api.NewMinimalVMI("testvm")
-		vm.Spec.Domain.Devices.Interfaces = []v1.Interface{{
-			Name:                   "foo",
-			InterfaceBindingMethod: v1.InterfaceBindingMethod{Bridge: &v1.InterfaceBridge{}},
-		}}
-		validator := admitter.NewValidator(k8sfield.NewPath("fake"), &vm.Spec, stubSlirpClusterConfigChecker{})
-		Expect(validator.Validate()).To(BeEmpty())
-	})
-
 	It("support only a single pod network", func() {
 		const net1Name = "default"
 		const net2Name = "default2"
